updated: move method down to class AutoLayout
