+++ src/Domain/Layout/AutoLayout.cpp
@@ -1,10 +1,13 @@
 #include "AutoLayout.hpp"
 
 #include "Log.h"
+#include "Rule.hpp"
 #include "View.hpp"
 
 #include <memory>
 #include <vector>
+
+using namespace VGG::Layout::Internal::Rule;
 
 namespace VGG
 {
@@ -15,6 +18,108 @@ namespace Layout
 
 namespace Internal
 {
+
+namespace
+{
+using LengthTypes = VGG::Layout::Internal::Rule::Length::Types;
+unit toLibUnit(LengthTypes type)
+{
+  switch (type)
+  {
+    case LengthTypes::px:
+      return unit_point;
+    case LengthTypes::percent:
+      return unit_percent;
+    case LengthTypes::fit_content:
+      return unit_auto;
+    default:
+      return unit_undefine;
+  }
+}
+
+position toLibPosition(Position::Types type)
+{
+  switch (type)
+  {
+    case Position::Types::Relative:
+      return position_relative;
+    case Position::Types::Absolute:
+      return position_absolute;
+
+    default: // todo
+      return position_absolute;
+  }
+}
+
+direction toLibDirecion(FlexboxLayout::Direction type)
+{
+  switch (type)
+  {
+    case FlexboxLayout::Direction::Horizontal:
+      return direction_row;
+
+    default:
+      return direction_column;
+  }
+}
+
+justify_content toLibJustifyContent(FlexboxLayout::JustifyContent type)
+{
+  switch (type)
+  {
+    case FlexboxLayout::JustifyContent::Start:
+      return justify_content_flex_start;
+    case FlexboxLayout::JustifyContent::Center:
+      return justify_content_center;
+    case FlexboxLayout::JustifyContent::End:
+      return justify_content_flex_end;
+    case FlexboxLayout::JustifyContent::SpaceBetween:
+      return justify_content_space_between;
+    case FlexboxLayout::JustifyContent::SpaceAround:
+      return justify_content_space_around;
+    case FlexboxLayout::JustifyContent::SpaceEvenly:
+      return justify_content_space_evenly;
+  }
+}
+
+align_content toLibAlignContent(AlignStyle type)
+{
+  switch (type)
+  {
+    case AlignStyle::Start:
+      return align_content_flex_start;
+    case AlignStyle::Center:
+      return align_content_center;
+    case AlignStyle::End:
+      return align_content_flex_end;
+  }
+}
+
+align_items toLibAlignItem(AlignStyle type)
+{
+  switch (type)
+  {
+    case AlignStyle::Start:
+      return align_items_flex_start;
+    case AlignStyle::Center:
+      return align_items_center;
+    case AlignStyle::End:
+      return align_items_flex_end;
+  }
+}
+
+wrap toLibWrap(FlexboxLayout::Wrap type)
+{
+  switch (type)
+  {
+    case FlexboxLayout::Wrap::NoWrap:
+      return wrap_no_wrap;
+    case FlexboxLayout::Wrap::Wrap:
+      return wrap_wrap;
+  }
+}
+
+} // namespace
 
 namespace
 {
@@ -257,6 +362,127 @@ flexbox_node* AutoLayout::getFlexNode()
 grid_layout* AutoLayout::getGridNode()
 {
   return m_gridNodePtr;
+}
+
+void AutoLayout::configure()
+{
+  auto sharedRule = rule.lock();
+  if (!sharedRule)
+  {
+    return;
+  }
+
+  auto sharedView = view.lock();
+  if (!sharedView)
+  {
+    return;
+  }
+
+  if (const auto detail = std::get_if<FlexboxLayout>(&sharedRule->layout))
+  {
+    DEBUG("LayoutView::configureAutoLayout, flex layout, view[%p, %s]",
+          sharedView.get(),
+          sharedView->path().c_str());
+
+    auto node = createFlexNode();
+    configureNode(node, sharedRule);
+
+    node->set_direction(toLibDirecion(detail->direction));
+    node->set_justify_content(toLibJustifyContent(detail->justify_content));
+    node->set_align_items(toLibAlignItem(detail->align_items));
+    node->set_align_content(toLibAlignContent(detail->align_content));
+    node->set_wrap(toLibWrap(detail->wrap));
+    node->set_gap(gap_row, detail->row_gap);
+    node->set_gap(gap_column, detail->column_gap);
+    node->set_padding(padding_top, detail->padding.top);
+    node->set_padding(padding_right, detail->padding.right);
+    node->set_padding(padding_bottom, detail->padding.bottom);
+    node->set_padding(padding_left, detail->padding.left);
+  }
+  else if (const auto detail = std::get_if<GridLayout>(&sharedRule->layout))
+  {
+    DEBUG("LayoutView::configureAutoLayout, grid layout, view[%p, %s]",
+          sharedView.get(),
+          sharedView->path().c_str());
+
+    // todo
+    // auto node = new grid_layout;
+    // autoLayout->gridNode.reset(node);
+  }
+
+  if (const auto detail = std::get_if<FlexboxItem>(&sharedRule->item_in_layout))
+  {
+    DEBUG("LayoutView::configureAutoLayout, flex item, view[%p, %s]",
+          sharedView.get(),
+          sharedView->path().c_str());
+
+    auto node = getFlexNode();
+    if (!node)
+    {
+      node = createFlexNode();
+    }
+    configureNode(node, sharedRule);
+
+    node->set_position(toLibPosition(detail->position.value));
+    node->set_grow(detail->flex_grow);
+
+    if (detail->top.has_value())
+    {
+      node->set_ltrb(ltrb_top, detail->top->value);
+    }
+    if (detail->right.has_value())
+    {
+      node->set_ltrb(ltrb_right, detail->right->value);
+    }
+    if (detail->bottom.has_value())
+    {
+      node->set_ltrb(ltrb_bottom, detail->bottom->value);
+    }
+    if (detail->left.has_value())
+    {
+      node->set_ltrb(ltrb_left, detail->left->value);
+    }
+  }
+  else if (const auto detail = std::get_if<GridItem>(&sharedRule->item_in_layout))
+  {
+    DEBUG("LayoutView::configureAutoLayout, grid item, view[%p, %s]",
+          sharedView.get(),
+          sharedView->path().c_str());
+    // todo
+  }
+}
+
+void AutoLayout::configureNode(flexbox_node* node,
+                               std::shared_ptr<VGG::Layout::Internal::Rule::Rule> sharedRule)
+{
+  node->set_width(toLibUnit(sharedRule->width.value.types), sharedRule->width.value.value);
+  if (sharedRule->max_width.has_value())
+  {
+    node->set_max_width(toLibUnit(sharedRule->max_width->value.types),
+                        sharedRule->max_width->value.value);
+  }
+  if (sharedRule->min_width.has_value())
+  {
+    node->set_min_width(toLibUnit(sharedRule->min_width->value.types),
+                        sharedRule->min_width->value.value);
+  }
+
+  node->set_height(toLibUnit(sharedRule->height.value.types), sharedRule->height.value.value);
+  if (sharedRule->max_height.has_value())
+  {
+    node->set_max_height(toLibUnit(sharedRule->max_height->value.types),
+                         sharedRule->max_height->value.value);
+  }
+  if (sharedRule->min_height.has_value())
+  {
+    node->set_min_height(toLibUnit(sharedRule->min_height->value.types),
+                         sharedRule->min_height->value.value);
+  }
+
+  if (sharedRule->aspect_ratio.has_value())
+  {
+    // todo
+  }
 }
 
 } // namespace Internal
